requirements.txt
anyio==3.6.1
asyncpg==0.26.0
atomicwrites==1.4.1
attrs==22.1.0
autopep8==2.0.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
fastapi==0.79.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.9.1
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.21.0
requests==2.28.1
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
